+++ src/database/structs.rs
@@ -98,22 +98,9 @@ pub enum PreferenceValue {
 pub struct RecordEvidence {
     pub player_login: String,
     pub map_uid: String,
+    pub nb_laps: i32,
     pub millis: i32,
     pub timestamp: NaiveDateTime,
-    pub sectors: Vec<RecordSector>,
-}
-
-/// Detailed checkpoint data recorded at the end of a sector.
-#[derive(Clone, Debug, PartialEq)]
-pub struct RecordSector {
-    /// First sector has index 0, and so on.
-    pub index: i32,
-
-    /// Total millis at time of crossing the checkpoint at this index.
-    pub cp_millis: i32,
-
-    /// Speed at time of crossing the checkpoint at this index.
-    pub cp_speed: f32,
 }
 
 /// Detailed record data, that is only missing speed & distance
@@ -123,12 +110,17 @@ pub struct Record {
     /// The UID of the map this record was set on.
     pub map_uid: String,
 
-    /// The player's map rank, which is the rank of this record
-    /// in the ranking of all records on this map.
-    pub map_rank: i64,
-
     /// The login of the player that has set this record.
     pub player_login: String,
+
+    /// The number of laps for this record.
+    ///
+    /// Use `0` if the map is not multi-lap, or to count flying lap records.
+    pub nb_laps: i32,
+
+    /// The player's map rank, which is the rank of this record
+    /// in the ranking of all records with the same lap count on this map.
+    pub map_rank: i64,
 
     /// The formatted display name of the player that has set this record.
     pub player_display_name: DisplayString,
@@ -138,12 +130,11 @@ pub struct Record {
 
     /// The moment this record was set.
     pub timestamp: NaiveDateTime,
-
-    /// Checkpoint data.
-    pub sectors: Vec<RecordSector>,
 }
 
 /// A rank of a player's record on a specific map.
+///
+/// On multi-lap maps, this is the
 #[derive(Debug)]
 pub struct MapRank {
     pub map_uid: String,